More cleanup + documentation
